Mms: sets endMmsConnectivity() as last operation
When a user receives an Mms and there is no more pending requests, the method
endMmsConnectivity() is called to turn the data connection off and clear the
wakelock. Sometimes the phone deep sleeps before the user gets Mms
notifications, so this patch makes endMmsConnectivity() to be the last
operation so users can be notified of Mms's.

Change-Id: Iff518acbf98486ede7f24e1439b021d26cdbccc1